+++ retrieval/targetmanager.go
@@ -323,107 +323,6 @@ func (ts *targetSet) update(name string, tgroup *config.TargetGroup) error {
 	return nil
 }
 
-// scrapePool manages scrapes for sets of targets.
-type scrapePool struct {
-	appender storage.SampleAppender
-
-	ctx     context.Context
-	mtx     sync.RWMutex
-	tgroups map[string]map[model.Fingerprint]*Target
-}
-
-func newScrapePool(app storage.SampleAppender) *scrapePool {
-	return &scrapePool{
-		appender: app,
-		tgroups:  map[string]map[model.Fingerprint]*Target{},
-	}
-}
-
-func (sp *scrapePool) stop() {
-	var wg sync.WaitGroup
-
-	sp.mtx.RLock()
-
-	for _, tgroup := range sp.tgroups {
-		for _, t := range tgroup {
-			wg.Add(1)
-
-			go func(t *Target) {
-				t.StopScraper()
-				wg.Done()
-			}(t)
-		}
-	}
-	sp.mtx.RUnlock()
-
-	wg.Wait()
-}
-
-func (sp *scrapePool) sync(tgroups map[string]map[model.Fingerprint]*Target) {
-	sp.mtx.Lock()
-
-	var (
-		wg         sync.WaitGroup
-		newTgroups = map[string]map[model.Fingerprint]*Target{}
-	)
-
-	for source, targets := range tgroups {
-		var (
-			prevTargets = sp.tgroups[source]
-			newTargets  = map[model.Fingerprint]*Target{}
-		)
-		newTgroups[source] = newTargets
-
-		for fp, tnew := range targets {
-			// If the same target existed before, we let it run and replace
-			// the new one with it.
-			if told, ok := prevTargets[fp]; ok {
-				newTargets[fp] = told
-			} else {
-				newTargets[fp] = tnew
-				go tnew.RunScraper(sp.appender)
-			}
-		}
-		for fp, told := range prevTargets {
-			// A previous target is no longer in the group.
-			if _, ok := targets[fp]; !ok {
-				wg.Add(1)
-
-				go func(told *Target) {
-					told.StopScraper()
-					wg.Done()
-				}(told)
-			}
-		}
-	}
-
-	// Stop scrapers for target groups that disappeared completely.
-	for source, targets := range sp.tgroups {
-		if _, ok := tgroups[source]; ok {
-			continue
-		}
-		for _, told := range targets {
-			wg.Add(1)
-
-			go func(told *Target) {
-				told.StopScraper()
-				wg.Done()
-			}(told)
-		}
-	}
-
-	sp.tgroups = newTgroups
-
-	// Wait for all potentially stopped scrapers to terminate.
-	// This covers the case of flapping targets. If the server is under high load, a new scraper
-	// may be active and tries to insert. The old scraper that didn't terminate yet could still
-	// be inserting a previous sample set.
-	wg.Wait()
-
-	// TODO(fabxc): maybe this can be released earlier with subsequent refactoring.
-	sp.mtx.Unlock()
-}
-
 // providersFromConfig returns all TargetProviders configured in cfg.
 func providersFromConfig(cfg *config.ScrapeConfig) map[string]TargetProvider {
 	providers := map[string]TargetProvider{}
